pyproject.toml
[project]
name = "guide"
version = "0.1.0"
description = "Declarative purpose reification system"
readme = "README.md"
authors = [
    { name = "xmkqv", email = "m@qxotk.ai" }
]
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0",
    "pytest>=8.0",
    "attrs>=25.0",
    "numpy>=2.0.0",
    "returns>=0.24.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "structlog>=24.0.0",
    "flet>=0.28.3",
    "fletxr>=0.1.3",
    "expression[pydantic]>=5.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest-reportlog>=0.4.0",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]

[project.scripts]
guide = "guide.app.__main__:main"

[tool.pytest.ini_options]
markers = [
    "mission: Mission validation tests",
]
addopts = "--report-log=results/results.jsonl"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
extend = "configs/ruff.toml"
